finish lambda to upload pdfs to s3
