+++ .serverless/cloudformation-template-update-stack.json
@@ -56,76 +56,76 @@
         }
       }
     },
-    "FetchUnderscorepapersLogGroup": {
+    "FetchUnderscoreentryUnderscoredataLogGroup": {
       "Type": "AWS::Logs::LogGroup",
       "Properties": {
-        "LogGroupName": "/aws/lambda/asp2-dev-fetch_papers"
+        "LogGroupName": "/aws/lambda/asp2-dev-fetch_entry_data"
       }
     },
-    "DownloadUnderscorepdfsLogGroup": {
+    "ProcessUnderscorestreamLogGroup": {
       "Type": "AWS::Logs::LogGroup",
       "Properties": {
-        "LogGroupName": "/aws/lambda/asp2-dev-download_pdfs"
+        "LogGroupName": "/aws/lambda/asp2-dev-process_stream"
       }
     },
-    "FetchUnderscorepapersLambdaFunction": {
+    "FetchUnderscoreentryUnderscoredataLambdaFunction": {
       "Type": "AWS::Lambda::Function",
       "Properties": {
         "Code": {
           "S3Bucket": {
             "Ref": "ServerlessDeploymentBucket"
           },
-          "S3Key": "serverless/asp2/dev/1604887608769-2020-11-09T02:06:48.769Z/asp2.zip"
+          "S3Key": "serverless/asp2/dev/1604998737254-2020-11-10T08:58:57.254Z/asp2.zip"
         },
-        "FunctionName": "asp2-dev-fetch_papers",
-        "Handler": "lambdas/fetch_papers.main",
+        "FunctionName": "asp2-dev-fetch_entry_data",
+        "Handler": "lambdas/fetch_entry_data.main",
         "MemorySize": 1024,
         "Role": "arn:aws:iam::555989344246:role/asp2_role",
         "Runtime": "python3.8",
         "Timeout": 6
       },
       "DependsOn": [
-        "FetchUnderscorepapersLogGroup"
+        "FetchUnderscoreentryUnderscoredataLogGroup"
       ]
     },
-    "FetchUnderscorepapersLambdaVersionkz7J4uEmyBnnD2XNw6m5Jrv3RN3ger0WMD7HucHQyU": {
+    "FetchUnderscoreentryUnderscoredataLambdaVersion6yB20hFwzkCxWNKsdRwFeajfzwkN1Wzi3vPmlisPmg": {
       "Type": "AWS::Lambda::Version",
       "DeletionPolicy": "Retain",
       "Properties": {
         "FunctionName": {
-          "Ref": "FetchUnderscorepapersLambdaFunction"
+          "Ref": "FetchUnderscoreentryUnderscoredataLambdaFunction"
         },
-        "CodeSha256": "pECWG1UszcB9OhyxwEzcc0Rg1gbDKgejaeyL/ZjcrPs="
+        "CodeSha256": "0k1ELvTrRG9PVQQ7ORKdy1sqeLudRHZDxYad7ymhG2w="
       }
     },
-    "DownloadUnderscorepdfsLambdaFunction": {
+    "ProcessUnderscorestreamLambdaFunction": {
       "Type": "AWS::Lambda::Function",
       "Properties": {
         "Code": {
           "S3Bucket": {
             "Ref": "ServerlessDeploymentBucket"
           },
-          "S3Key": "serverless/asp2/dev/1604887608769-2020-11-09T02:06:48.769Z/asp2.zip"
+          "S3Key": "serverless/asp2/dev/1604998737254-2020-11-10T08:58:57.254Z/asp2.zip"
         },
-        "FunctionName": "asp2-dev-download_pdfs",
-        "Handler": "lambdas/download_pdfs.main",
+        "FunctionName": "asp2-dev-process_stream",
+        "Handler": "lambdas/process_stream.main",
         "MemorySize": 1024,
         "Role": "arn:aws:iam::555989344246:role/asp2_role",
         "Runtime": "python3.8",
         "Timeout": 6
       },
       "DependsOn": [
-        "DownloadUnderscorepdfsLogGroup"
+        "ProcessUnderscorestreamLogGroup"
       ]
     },
-    "DownloadUnderscorepdfsLambdaVersionDKushMeZHYz7KRSFOqXXKiLhzTJ8grUDuPvGnBW1zs": {
+    "ProcessUnderscorestreamLambdaVersionM0x6omTn64dEgXulnOU3UuVnbJ3JltTGkUROfFkhXM": {
       "Type": "AWS::Lambda::Version",
       "DeletionPolicy": "Retain",
       "Properties": {
         "FunctionName": {
-          "Ref": "DownloadUnderscorepdfsLambdaFunction"
+          "Ref": "ProcessUnderscorestreamLambdaFunction"
         },
-        "CodeSha256": "pECWG1UszcB9OhyxwEzcc0Rg1gbDKgejaeyL/ZjcrPs="
+        "CodeSha256": "0k1ELvTrRG9PVQQ7ORKdy1sqeLudRHZDxYad7ymhG2w="
       }
     }
   },
@@ -135,16 +135,16 @@
         "Ref": "ServerlessDeploymentBucket"
       }
     },
-    "FetchUnderscorepapersLambdaFunctionQualifiedArn": {
+    "FetchUnderscoreentryUnderscoredataLambdaFunctionQualifiedArn": {
       "Description": "Current Lambda function version",
       "Value": {
-        "Ref": "FetchUnderscorepapersLambdaVersionkz7J4uEmyBnnD2XNw6m5Jrv3RN3ger0WMD7HucHQyU"
+        "Ref": "FetchUnderscoreentryUnderscoredataLambdaVersion6yB20hFwzkCxWNKsdRwFeajfzwkN1Wzi3vPmlisPmg"
       }
     },
-    "DownloadUnderscorepdfsLambdaFunctionQualifiedArn": {
+    "ProcessUnderscorestreamLambdaFunctionQualifiedArn": {
       "Description": "Current Lambda function version",
       "Value": {
-        "Ref": "DownloadUnderscorepdfsLambdaVersionDKushMeZHYz7KRSFOqXXKiLhzTJ8grUDuPvGnBW1zs"
+        "Ref": "ProcessUnderscorestreamLambdaVersionM0x6omTn64dEgXulnOU3UuVnbJ3JltTGkUROfFkhXM"
       }
     }
   }
