--- example_events/download_pdfs.json
@@ -1,70 +1,0 @@
-{
-  "Records": [
-    {
-      "eventID": "583d7a1aecc801727f489fd9764e3eba",
-      "eventName": "INSERT",
-      "eventVersion": "1.1",
-      "eventSource": "aws:dynamodb",
-      "awsRegion": "us-east-2",
-      "dynamodb": {
-        "ApproximateCreationDateTime": 1604701128.0,
-        "Keys": {
-          "rawid": { "S": "2011.02952" }
-        },
-        "NewImage": {
-          "rawid": {
-            "S": "2011.02952"
-          },
-          "links": {
-            "L": [
-              {
-                "M": {
-                  "rel": {
-                    "S": "alternate"
-                  },
-                  "href": {
-                    "S": "http://arxiv.org/abs/2011.02952v1"
-                  },
-                  "type": {
-                    "S": "text/html"
-                  }
-                }
-              },
-              {
-                "M": {
-                  "rel": {
-                    "S": "related"
-                  },
-                  "href": {
-                    "S": "http://arxiv.org/pdf/2011.02952v1"
-                  },
-                  "title": {
-                    "S": "pdf"
-                  },
-                  "type": {
-                    "S": "application/pdf"
-                  }
-                }
-              }
-            ]
-          },
-          "version": { "N": "1" },
-          "authors": {
-            "L": [
-              { "M": { "name": { "S": "Sebastian Buschjäger" } } },
-              { "M": { "name": { "S": "Lukas Pfahler" } } },
-              { "M": { "name": { "S": "Katharina Morik" } } }
-            ]
-          },
-          "title": {
-            "S": "Generalized Negative Correlation Learning for Deep Ensembling"
-          }
-        },
-        "SequenceNumber": "12247700000000005054456435",
-        "SizeBytes": 284,
-        "StreamViewType": "NEW_IMAGE"
-      },
-      "eventSourceARN": "arn:aws:dynamodb:us-east-2:555989344246:table/asp2-fetch-results/stream/2020-11-06T21:53:27.159"
-    }
-  ]
-}
