--- example_events/fetch_papers.json
@@ -1,8 +1,0 @@
-{
-    "search_query": "cat:stat.ML",
-    "start_index": 0,
-    "max_index": 100,
-    "res_per_iter": 10,
-    "wait_time": 0.0,
-    "break_on_no_add": 1
-}
